+++ include/wx/gtk/notebook.h
@@ -56,17 +56,17 @@ public:
     // set the currently selected page, return the index of the previously
     // selected one (or -1 on error)
     // NB: this function will _not_ generate wxEVT_NOTEBOOK_PAGE_xxx events
-  int SetSelection(int nPage);
+  int SetSelection(size_t nPage);
     // get the currently selected page
   int GetSelection() const;
 
     // set/get the title of a page
-  bool SetPageText(int nPage, const wxString& strText);
-  wxString GetPageText(int nPage) const;
+  bool SetPageText(size_t nPage, const wxString& strText);
+  wxString GetPageText(size_t nPage) const;
 
     // sets/returns item's image index in the current image list
-  int  GetPageImage(int nPage) const;
-  bool SetPageImage(int nPage, int nImage);
+  int  GetPageImage(size_t nPage) const;
+  bool SetPageImage(size_t nPage, int nImage);
 
   // control the appearance of the notebook pages
     // set the size (the same for all pages)
@@ -79,7 +79,7 @@ public:
   // operations
   // ----------
     // remove one page from the notebook
-  bool DeletePage(int nPage);
+  bool DeletePage(size_t nPage);
     // remove all pages
   bool DeleteAllPages();
 
@@ -127,7 +127,7 @@ public:
 
 protected:
     // remove one page from the notebook but do not destroy it
-    virtual wxNotebookPage *DoRemovePage(int nPage);
+    virtual wxNotebookPage *DoRemovePage(size_t nPage);
 
 private:
     // the padding set by SetPadding()
